+++ .mocharc.cjs
@@ -2,12 +2,12 @@ module.exports = {
   loader: 'ts-node/esm',
   extension: ['ts'],
   recursive: true,
-  timeout: 5000,
+  timeout: 10000,
   // Write JUnit XML report when we're running in CI
-  ...(process.env.CI
-    ? {
-        reporter: 'mocha-junit-reporter',
-        'reporter-option': `mochaFile=./tmp/tests/junit-${Date.now()}.xml`,
-      }
-    : {}),
+  // ...(process.env.CI
+  //   ? {
+  //       reporter: 'mocha-junit-reporter',
+  //       'reporter-option': `mochaFile=./tmp/tests/junit-${Date.now()}.xml`,
+  //     }
+  //   : {}),
 }