+++ src/bin/one-report-publisher.ts
@@ -10,9 +10,9 @@ program.requiredOption(
   process.env.ONE_REPORT_URL
 )
 program.requiredOption(
-  '-o, --organization <id>',
-  'OneReport organization id. Defaults to $ONE_REPORT_ORGANIZATION',
-  process.env.ONE_REPORT_ORGANIZATION
+  '-p, --project-id <id>',
+  'OneReport project id. Defaults to $ONE_REPORT_PROJECT_ID',
+  process.env.ONE_REPORT_PROJECT_ID
 )
 program.requiredOption(
   '-t, --token <token>',
@@ -26,7 +26,7 @@ program.option('--no-zip', 'Do not zip non .zip files', false)
 
 program.parse(process.argv)
 const {
-  organization,
+  projectId,
   token,
   reports: globs,
   maxTime,
@@ -40,7 +40,7 @@ async function main() {
   const responseBodies = await publish<OneReportResponseBody>(
     globs,
     !noZip,
-    organization,
+    projectId,
     baseUrl,
     process.env,
     tokenAuthenticator(token),